fix(app): action_args and result from api on execution tab
