+++ apps/app/src/components/app/Transactions/Receipt.tsx
@@ -7,10 +7,12 @@ import ErrorMessage from '@/components/app/common/ErrorMessage';
 import FaHourglassStart from '@/components/app/Icons/FaHourglassStart';
 import FileSlash from '@/components/app/Icons/FileSlash';
 import ReceiptRow from '@/components/app/Transactions/Receipts/ReceiptRow';
+import { mapRpcActionToAction } from '@/utils/app/near';
+import { useEffect, useState } from 'react';
+import { isEmpty } from 'lodash';
 
 interface Props {
   hash: string;
-  receipt: null | ReceiptsPropsInfo;
   rpcTxn: RPCTransactionInfo;
   statsData: {
     stats: Array<{
@@ -22,14 +24,77 @@ interface Props {
 }
 
 const Receipt = (props: Props) => {
-  const {
-    hash,
-    receipt: rpcReceipt,
-    statsData,
-    txn,
-    rpcTxn,
-    apiTxnActionsData,
-  } = props;
+  const { hash, statsData, txn, rpcTxn, apiTxnActionsData } = props;
+
+  const [rpcReceipt, setReceipt] = useState<null | ReceiptsPropsInfo>(null);
+
+  function transactionReceipts(txn: RPCTransactionInfo) {
+    const actions: any =
+      txn?.transaction?.actions &&
+      txn?.transaction?.actions?.map((txn) => mapRpcActionToAction(txn));
+    const receipts = txn?.receipts;
+    const receiptsOutcome = txn?.receipts_outcome;
+
+    if (
+      receipts?.length === 0 ||
+      receipts?.[0]?.receipt_id !== receiptsOutcome?.[0]?.id
+    ) {
+      receipts?.unshift({
+        predecessor_id: txn?.transaction?.signer_id,
+        receipt: actions,
+        receipt_id: receiptsOutcome?.[0]?.id,
+        receiver_id: txn?.transaction?.receiver_id,
+      });
+    }
+
+    const receiptOutcomesByIdMap = new Map();
+    const receiptsByIdMap = new Map();
+
+    receiptsOutcome &&
+      receiptsOutcome?.forEach((receipt) => {
+        receiptOutcomesByIdMap?.set(receipt?.id, receipt);
+      });
+
+    receipts &&
+      receipts?.forEach((receiptItem) => {
+        receiptsByIdMap?.set(receiptItem?.receipt_id, {
+          ...receiptItem,
+          actions:
+            receiptItem?.receipt_id === receiptsOutcome?.[0]?.id
+              ? actions
+              : receiptItem?.receipt?.Action?.actions &&
+                receiptItem?.receipt?.Action?.actions.map((receipt) =>
+                  mapRpcActionToAction(receipt),
+                ),
+        });
+      });
+
+    const collectReceipts = (receiptHash: any) => {
+      const receipt = receiptsByIdMap?.get(receiptHash);
+      const receiptOutcome = receiptOutcomesByIdMap?.get(receiptHash);
+
+      return {
+        ...receipt,
+        ...receiptOutcome,
+        outcome: {
+          ...receiptOutcome?.outcome,
+          outgoing_receipts:
+            receiptOutcome?.outcome?.receipt_ids &&
+            receiptOutcome?.outcome?.receipt_ids?.map(collectReceipts),
+        },
+      };
+    };
+
+    return collectReceipts(receiptsOutcome?.[0]?.id);
+  }
+
+  useEffect(() => {
+    if (!isEmpty(rpcTxn)) {
+      setReceipt(transactionReceipts(rpcTxn));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [rpcTxn]);
+
   const receipt = !apiTxnActionsData?.receiptData
     ? rpcReceipt
     : apiTxnActionsData?.receiptData;
@@ -64,11 +129,7 @@ const Receipt = (props: Props) => {
               </div>
             </div>
           ) : (
-            <ReceiptRow
-              receipt={receipt}
-              statsData={statsData}
-              rpcReceipt={rpcReceipt}
-            />
+            <ReceiptRow receipt={receipt} statsData={statsData} />
           )}
         </div>
       )}
