+++ apps/app/src/components/app/Transactions/Receipts/ReceiptStatus.tsx
@@ -52,7 +52,7 @@ const ReceiptStatus = (props: ReceiptStatsProps) => {
     return (
       <textarea
         className="block appearance-none outline-none w-full border rounded-lg bg-gray-100 dark:bg-black-200 dark:border-black-200 p-3 mt-3 resize-y"
-        defaultValue={JSON.stringify(status.Failure, null, 2)}
+        defaultValue={JSON.stringify(status?.Failure?.error_message, null, 2)}
         readOnly
         rows={4}
       ></textarea>
