+++ apps/app/src/components/app/Transactions/TxnsTabActions.tsx
@@ -1,6 +1,5 @@
 'use client';
 import classNames from 'classnames';
-import { isEmpty } from 'lodash';
 import { useRouter } from 'next/navigation';
 import { useEffect, useMemo, useRef, useState } from 'react';
 
@@ -14,12 +13,7 @@ import {
   calculateTotalGas,
   txnFee,
 } from '@/utils/near';
-import { mapRpcActionToAction } from '@/utils/near';
-import {
-  ExecutionOutcomeWithIdView,
-  ReceiptsPropsInfo,
-  RPCTransactionInfo,
-} from '@/utils/types';
+import { ExecutionOutcomeWithIdView } from '@/utils/types';
 
 import ErrorMessage from '@/components/app/common/ErrorMessage';
 import FileSlash from '@/components/app/Icons/FileSlash';
@@ -53,74 +47,6 @@ const TxnsTabActions = ({
   const retryCount = useRef(0);
   const timeoutRef = useRef<NodeJS.Timeout | null>(null);
   const router = useRouter();
-  const [receipt, setReceipt] = useState<null | ReceiptsPropsInfo>(null);
-
-  function transactionReceipts(txn: RPCTransactionInfo) {
-    const actions: any =
-      txn?.transaction?.actions &&
-      txn?.transaction?.actions?.map((txn) => mapRpcActionToAction(txn));
-    const receipts = txn?.receipts;
-    const receiptsOutcome = txn?.receipts_outcome;
-
-    if (
-      receipts?.length === 0 ||
-      receipts?.[0]?.receipt_id !== receiptsOutcome?.[0]?.id
-    ) {
-      receipts?.unshift({
-        predecessor_id: txn?.transaction?.signer_id,
-        receipt: actions,
-        receipt_id: receiptsOutcome?.[0]?.id,
-        receiver_id: txn?.transaction?.receiver_id,
-      });
-    }
-
-    const receiptOutcomesByIdMap = new Map();
-    const receiptsByIdMap = new Map();
-
-    receiptsOutcome &&
-      receiptsOutcome?.forEach((receipt) => {
-        receiptOutcomesByIdMap?.set(receipt?.id, receipt);
-      });
-
-    receipts &&
-      receipts?.forEach((receiptItem) => {
-        receiptsByIdMap?.set(receiptItem?.receipt_id, {
-          ...receiptItem,
-          actions:
-            receiptItem?.receipt_id === receiptsOutcome?.[0]?.id
-              ? actions
-              : receiptItem?.receipt?.Action?.actions &&
-                receiptItem?.receipt?.Action?.actions.map((receipt) =>
-                  mapRpcActionToAction(receipt),
-                ),
-        });
-      });
-
-    const collectReceipts = (receiptHash: any) => {
-      const receipt = receiptsByIdMap?.get(receiptHash);
-      const receiptOutcome = receiptOutcomesByIdMap?.get(receiptHash);
-
-      return {
-        ...receipt,
-        ...receiptOutcome,
-        outcome: {
-          ...receiptOutcome?.outcome,
-          outgoing_receipts:
-            receiptOutcome?.outcome?.receipt_ids &&
-            receiptOutcome?.outcome?.receipt_ids?.map(collectReceipts),
-        },
-      };
-    };
-
-    return collectReceipts(receiptsOutcome?.[0]?.id);
-  }
-
-  useEffect(() => {
-    if (!isEmpty(rpcTxn)) {
-      setReceipt(transactionReceipts(rpcTxn));
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [rpcTxn]);
 
   const useRpcStoreWithProviders = () => {
     const setProviders = useRpcStore((state) => state.setProviders);
@@ -351,7 +277,6 @@ const TxnsTabActions = ({
                 {tab === 'execution' && (
                   <Receipt
                     hash={hash}
-                    receipt={receipt}
                     rpcTxn={rpcTxn}
                     statsData={stats}
                     txn={txn ? txn : rpcData}
